Fix DataFrame.from_records test
Currently `test_from_records_to_records` specifies a misleading error message when the shape of the input data does not match the shape implied by the indices.  In particular, the error message omits the offending dimension entirely!

I can see how this error made it into the code, but how on earth did this behavior get enshrined in a test?